+++ oghma_core/include/hard_limit.h
@@ -1,10 +1,8 @@
-// 
-// General-purpose Photovoltaic Device Model gpvdm.com - a drift diffusion
-// base/Shockley-Read-Hall model for 1st, 2nd and 3rd generation solarcells.
-// The model can simulate OLEDs, Perovskite cells, and OFETs.
-// 
-// Copyright 2008-2022 Roderick C. I. MacKenzie https://www.gpvdm.com
-// r.c.i.mackenzie at googlemail.com
+//
+// OghmaNano - Organic and hybrid Material Nano Simulation tool
+// Copyright (C) 2008-2022 Roderick C. I. MacKenzie r.c.i.mackenzie at googlemail.com
+//
+// https://www.oghma-nano.com
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a
 // copy of this software and associated documentation files (the "Software"),
@@ -32,12 +30,12 @@
 
 #ifndef hard_limit_h
 #define hard_limit_h
-
+#include <g_io.h>
 #include <hard_limit_struct.h>
 #include <sim_struct.h>
 #include <json.h>
 
-void hard_limit_do(struct simulation *sim,char *token,long double *value);
+void hard_limit_do(struct simulation *sim,char *token,gdouble *value);
 void hard_limit_init(struct simulation *sim,struct hard_limit *hl);
 void hard_limit_free(struct simulation *sim,struct hard_limit *hl);
 void hard_limit_load(struct simulation *sim,struct hard_limit *hl,struct json_obj *json_hl);
