+++ oghma_core/include/math_2d.h
@@ -1,10 +1,8 @@
-// 
-// General-purpose Photovoltaic Device Model gpvdm.com - a drift diffusion
-// base/Shockley-Read-Hall model for 1st, 2nd and 3rd generation solarcells.
-// The model can simulate OLEDs, Perovskite cells, and OFETs.
-// 
-// Copyright 2008-2022 Roderick C. I. MacKenzie https://www.gpvdm.com
-// r.c.i.mackenzie at googlemail.com
+//
+// OghmaNano - Organic and hybrid Material Nano Simulation tool
+// Copyright (C) 2008-2022 Roderick C. I. MacKenzie r.c.i.mackenzie at googlemail.com
+//
+// https://www.oghma-nano.com
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a
 // copy of this software and associated documentation files (the "Software"),
@@ -25,12 +23,12 @@
 // SOFTWARE.
 // 
 
-
 /** @file math_2d.h
 	@brief Header file for math_2d.c
 */
 #ifndef math_2d_h
 #define math_2d_h
+#include <g_io.h>
 #include "advmath.h"
 #include <sim_struct.h>
 #include <png_image.h>
@@ -38,28 +36,28 @@
 
 struct math_2d
 {
-	long double **data;
+	double **data;
 	int x_len;
 	int y_len;
-	long double *y_mesh;
-	long double *x_mesh;
+	double *y_mesh;
+	double *x_mesh;
 
 };
 
 void math_2d_init(struct math_2d *in);
 void math_2d_malloc(struct math_2d *in, int x_len, int y_len);
-void math_2d_set_value(struct math_2d *in,long double value);
+void math_2d_set_value(struct math_2d *in,gdouble value);
 void math_2d_free(struct math_2d *in);
-void math_2d_init_y_mesh(struct math_2d *in,long double start, long double stop);
+void math_2d_init_y_mesh(struct math_2d *in,gdouble start, gdouble stop);
 void math_2d_save(struct simulation *sim,char *file_name,struct math_2d *data);
 void math_2d_to_triangles_simple(struct simulation *sim,struct triangles *tri,struct math_2d *in, int x_steps, int y_steps);
 void math_2d_to_triangles_front(struct simulation *sim,struct triangles *tri,struct math_2d *in, int x_steps, int y_steps);
 
 //math
-long double math_2d_y_min(struct simulation *sim,struct math_2d *in);
-long double math_2d_y_max(struct simulation *sim,struct math_2d *in);
+gdouble math_2d_y_min(struct simulation *sim,struct math_2d *in);
+gdouble math_2d_y_max(struct simulation *sim,struct math_2d *in);
 void math_2d_norm(struct simulation *sim,struct math_2d *in);
-void math_2d_add_long_double(struct simulation *sim,struct math_2d *in,long double val);
+void math_2d_add_long_double(struct simulation *sim,struct math_2d *in,gdouble val);
 
 //png
 void math_2d_png_load(struct simulation *sim,struct png_image *image,char* file_name);
