+++ oghma_core/include/solver.h
@@ -1,10 +1,8 @@
-// 
-// General-purpose Photovoltaic Device Model gpvdm.com - a drift diffusion
-// base/Shockley-Read-Hall model for 1st, 2nd and 3rd generation solarcells.
-// The model can simulate OLEDs, Perovskite cells, and OFETs.
-// 
-// Copyright 2008-2022 Roderick C. I. MacKenzie https://www.gpvdm.com
-// r.c.i.mackenzie at googlemail.com
+//
+// OghmaNano - Organic and hybrid Material Nano Simulation tool
+// Copyright (C) 2008-2022 Roderick C. I. MacKenzie r.c.i.mackenzie at googlemail.com
+//
+// https://www.oghma-nano.com
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a
 // copy of this software and associated documentation files (the "Software"),
@@ -32,6 +30,7 @@
 
 #ifndef solver_h
 #define solver_h
+#include <g_io.h>
 #include <sim_struct.h>
 
 void solver_test();
@@ -39,6 +38,6 @@ void set_solver_dump_every_matrix(int dump);
 void solver_precon(int col,int nz,int *Ti,int *Tj, double *Tx,double *b);
 int solver(struct simulation *sim,int col,int nz,int *Ti,int *Tj, double *Tx,double *b);
 void solver_dump_matrix(int col,int nz,int *Ti,int *Tj, double *Tx,double *b,char *index);
-void solver_dump_matrix_ld(int col,int nz,char **Tdebug,int *Ti,int *Tj, long double *Tx,long double *b,char *index);
+void solver_dump_matrix_ld(int col,int nz,char **Tdebug,int *Ti,int *Tj, gdouble *Tx,gdouble *b,char *index);
 void solver_print_time();
 #endif
