+++ oghma_core/plugins/complex_umfpack/solver.c
@@ -1,10 +1,8 @@
-// 
-// General-purpose Photovoltaic Device Model gpvdm.com - a drift diffusion
-// base/Shockley-Read-Hall model for 1st, 2nd and 3rd generation solarcells.
-// The model can simulate OLEDs, Perovskite cells, and OFETs.
-// 
-// Copyright 2008-2022 Roderick C. I. MacKenzie https://www.gpvdm.com
-// r.c.i.mackenzie at googlemail.com
+//
+// OghmaNano - Organic and hybrid Material Nano Simulation tool
+// Copyright (C) 2008-2022 Roderick C. I. MacKenzie r.c.i.mackenzie at googlemail.com
+//
+// https://www.oghma-nano.com
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a
 // copy of this software and associated documentation files (the "Software"),
@@ -43,7 +41,7 @@ EXPORT void dll_complex_matrix_init(struct matrix_solver_memory *msm)
 }
 
 
-EXPORT void dll_complex_matrix_solve(struct matrix_solver_memory *msm,int c_col,int c_nz,int *c_Ti,int *c_Tj, long double *c_Tx,long double *c_Txz,long double *c_b,long double *c_bz)
+EXPORT void dll_complex_matrix_solve(struct matrix_solver_memory *msm,int c_col,int c_nz,int *c_Ti,int *c_Tj, gdouble *c_Tx,gdouble *c_Txz,gdouble *c_b,gdouble *c_bz)
 {
 	complex_umfpack_solver(msm,c_col,c_nz,c_Ti,c_Tj, c_Tx,c_Txz,c_b,c_bz);
 }
